♻️ refactor(compute): split anomaly.rs into orchestration and signal modules (#92)
## Why (User Story)
As a developer, I want anomaly detection logic separated by concern so that
population statistics, individual signals, and multi-signal scoring are
independently maintainable.

## What Changed
- Split 662-line anomaly.rs into 3-file module:
  - mod.rs: Multi-signal orchestration, ClusterProvider trait, z-score scoring
  - population.rs: Population-level stats (mean, variance, std-dev)
  - signals.rs: Individual signal scorers (statistical, DBSCAN, behavioral, graph)

Related to #98
